Sortie des fonctions des hooks React Query vers un hook useDatabase et typage
