+++ web/src/lib/hooks/useWheelEntries.ts
@@ -1,30 +1,12 @@
 import { useQuery } from "@tanstack/react-query";
 import { QUERY_KEYS } from "@/lib/QueryKeys.ts";
-import { databases, Query } from "@/lib/appwrite.ts";
-import { databaseId, wheelCollectionId } from "@/lib/consts.ts";
+import { useDatabase } from "@/lib/hooks/useDatabase.ts";
 
 export const useWheelEntries = () => {
+  const { getWheelEntries } = useDatabase();
+
   return useQuery({
     queryKey: [QUERY_KEYS.WHEEL_ENTRIES],
-    queryFn: async () => {
-      const docs = await databases.listDocuments(
-        databaseId,
-        wheelCollectionId,
-        [Query.equal("disabled", false)],
-      );
-
-      // sort using the order field
-      docs.documents.sort((a, b) => {
-        console.log(a.order, b.order);
-        return a.order - b.order;
-      });
-
-      console.log(docs.documents.map((doc) => doc.name));
-
-      return docs.documents.map((doc) => ({
-        option: doc.name,
-        probability: doc.probability,
-      }));
-    },
+    queryFn: () => getWheelEntries(),
   });
 };
